store/quota query params: refactor min/default/max values, proper clamping of values from url query, bugfixing
