+++ src/packages/next/components/store/boost.tsx
@@ -29,13 +29,14 @@ import { computeCost } from "./compute-cost";
 import { InfoBar } from "./cost-info-bar";
 import { MemberHostingAndIdleTimeout } from "./member-idletime";
 import { QuotaConfig } from "./quota-config";
+import { decodeFormValues, encodeFormValues } from "./quota-query-params";
 import { Reset } from "./reset";
 import { RunLimit } from "./run-limit";
 import { SignInToPurchase } from "./sign-in-to-purchase";
 import { TitleDescription } from "./title-description";
 import { ToggleExplanations } from "./toggle-explanations";
 import { UsageAndDuration } from "./usage-and-duration";
-import { decodeFormValues, encodeFormValues, getType } from "./util";
+import { getType } from "./util";
 
 const { Text, Paragraph } = Typography;
