+++ src/packages/next/components/store/quota-config.tsx
@@ -6,14 +6,17 @@
 import { Icon } from "@cocalc/frontend/components/icon";
 import { displaySiteLicense } from "@cocalc/util/consts/site-license";
 import { plural } from "@cocalc/util/misc";
+import {
+  BOOST,
+  DISK_DEFAULT_GB,
+  MAX_RAM_GB,
+  REGULAR,
+} from "@cocalc/util/upgrades/consts";
 import { Col, Divider, Form, Radio, Row, Space, Tabs, Typography } from "antd";
 import A from "components/misc/A";
 import IntegerSlider from "components/misc/integer-slider";
-import { upgrades } from "@cocalc/util/upgrade-spec";
 import { Preset, PRESETS, Presets } from "./quota-config-presets";
 
-export const MAX_DISK = 15;
-export const MAX_GB_RAM = upgrades.max_per_project.memory / 1000;
 const { Text } = Typography;
 const { TabPane } = Tabs;
 
@@ -57,10 +60,7 @@ export const QuotaConfig: React.FC<Props> = (props: Props) => {
     }
   }
 
-  const min = boost ? 0 : 1;
-
-  // e.g. since we can't go beyond the max cpu, but the base license already provides one, don't let users select the max
-  const adjMax = boost ? 1 : 0;
+  const PARAMS = boost ? BOOST : REGULAR;
 
   function explainRam() {
     if (!showExplanations) return;
@@ -69,7 +69,7 @@ export const QuotaConfig: React.FC<Props> = (props: Props) => {
         This quota limits the total amount of memory a project can use. Note
         that RAM may be limited, if many other users are using the same host –
         though member hosting significantly reduces competition for RAM. We
-        recommend at least 2G! Beyond the overall maximum of {MAX_GB_RAM}G, we
+        recommend at least 2G! Beyond the overall maximum of {MAX_RAM_GB}G, we
         also offer{" "}
         <A href={"/store/dedicated?type=vm"}>dedicated virtual machines</A> with
         larger memory options.
@@ -85,20 +85,17 @@ export const QuotaConfig: React.FC<Props> = (props: Props) => {
   }
 
   function ram() {
-    const defaultRam = 2; // 2gb highly recommended
-    const maxRam = MAX_GB_RAM - adjMax;
-
     return (
       <Form.Item
         label="Shared RAM"
         name="ram"
-        initialValue={boost ? 0 : defaultRam}
+        initialValue={PARAMS.ram.dflt}
         extra={explainRam()}
       >
         <IntegerSlider
           disabled={disabled}
-          min={min}
-          max={maxRam}
+          min={PARAMS.ram.min}
+          max={PARAMS.ram.max}
           onChange={(ram) => {
             form.setFieldsValue({ ram });
             presetWasAdjusted();
@@ -116,7 +113,7 @@ export const QuotaConfig: React.FC<Props> = (props: Props) => {
       <Form.Item
         label="Shared CPUs"
         name="cpu"
-        initialValue={min}
+        initialValue={PARAMS.cpu.dflt}
         extra={
           showExplanations ? (
             <>
@@ -136,8 +133,8 @@ export const QuotaConfig: React.FC<Props> = (props: Props) => {
       >
         <IntegerSlider
           disabled={disabled}
-          min={min}
-          max={3 - adjMax}
+          min={PARAMS.cpu.min}
+          max={PARAMS.cpu.max}
           onChange={(cpu) => {
             form.setFieldsValue({ cpu });
             presetWasAdjusted();
@@ -152,22 +149,18 @@ export const QuotaConfig: React.FC<Props> = (props: Props) => {
 
   function disk() {
     // 2022-06: price increase "version 2": minimum disk we sell (also the free quota) is 3gb, not 1gb
-    const defaultDisk = 3;
-    const minDisk = boost ? 0 : defaultDisk;
-    const maxDisk = MAX_DISK - adjMax * defaultDisk;
-
     return (
       <Form.Item
         label="Disk space"
         name="disk"
-        initialValue={minDisk}
+        initialValue={PARAMS.disk.dflt}
         extra={
           showExplanations ? (
             <>
               Extra disk space lets you store a larger number of files.
               Snapshots and file edit history is included at no additional
-              charge. Each project receives at least {defaultDisk}G of storage
-              space. We also offer much larger{" "}
+              charge. Each project receives at least {DISK_DEFAULT_GB}G of
+              storage space. We also offer much larger{" "}
               <A href={"/store/dedicated?type=disk"}>dedicated disks</A>.
             </>
           ) : undefined
@@ -175,15 +168,17 @@ export const QuotaConfig: React.FC<Props> = (props: Props) => {
       >
         <IntegerSlider
           disabled={disabled}
-          min={minDisk}
-          max={maxDisk}
+          min={PARAMS.disk.min}
+          max={PARAMS.disk.max}
           onChange={(disk) => {
             form.setFieldsValue({ disk });
             presetWasAdjusted();
             onChange();
           }}
           units={"G Disk"}
-          presets={boost ? [0, 3, 6, maxDisk] : [3, 5, 10, maxDisk]}
+          presets={
+            boost ? [0, 3, 6, PARAMS.disk.max] : [3, 5, 10, PARAMS.disk.max]
+          }
         />
       </Form.Item>
     );
