+++ src/packages/next/components/store/site-license.tsx
@@ -26,13 +26,13 @@ import { InfoBar } from "./cost-info-bar";
 import { MemberHostingAndIdleTimeout } from "./member-idletime";
 import { QuotaConfig } from "./quota-config";
 import { PRESETS, Presets } from "./quota-config-presets";
+import { decodeFormValues, encodeFormValues } from "./quota-query-params";
 import { Reset } from "./reset";
 import { RunLimit } from "./run-limit";
 import { SignInToPurchase } from "./sign-in-to-purchase";
 import { TitleDescription } from "./title-description";
 import { ToggleExplanations } from "./toggle-explanations";
 import { UsageAndDuration } from "./usage-and-duration";
-import { decodeFormValues, encodeFormValues } from "./util";
 
 const STYLE: React.CSSProperties = {
   marginTop: "15px",
