+++ src/packages/next/components/store/util.ts
@@ -10,30 +10,13 @@ import {
 } from "@cocalc/frontend/misc/local-storage";
 import { ONE_DAY_MS } from "@cocalc/util/consts/billing";
 import { isValidUUID } from "@cocalc/util/misc";
-import {
-  endOfDay,
-  getDays,
-  roundToMidnight,
-  startOfDay,
-} from "@cocalc/util/stripe/timecalcs";
+import { endOfDay, getDays, startOfDay } from "@cocalc/util/stripe/timecalcs";
 import { DateRange } from "@cocalc/util/upgrades/shopping";
 import useCustomize from "lib/use-customize";
-import { isDate } from "lodash";
 import moment from "moment";
 import { NextRouter } from "next/router";
 import { useEffect, useMemo, useState } from "react";
 import { LicenseTypeInForms } from "./add-to-cart";
-import { MAX_DISK, MAX_GB_RAM } from "./quota-config";
-import { MAX_ALLOWED_RUN_LIMIT } from "./run-limit";
-import {
-  DEDICATED_DISK_SPEEDS,
-  DEDICATED_DISK_SIZES,
-  DEFAULT_DEDICATED_DISK_SIZE,
-  DEFAULT_DEDICATED_DISK_SPEED,
-  PRICES,
-  DEFAULT_DEDICATED_VM_MACHINE,
-} from "@cocalc/util/upgrades/dedicated";
-import { testDedicatedDiskNameBasic } from "@cocalc/util/licenses/check-disk-name-basics";
 
 // site license type in a form, we have 4 forms hence 4 types
 // later, this will be mapped to just "LicenseType", where boost and regular
@@ -187,221 +170,4 @@ export function useLicenseProject(router: NextRouter) {
     upgradeProjectDelete,
     storeLicenseProjectAssociation,
   };
-}
-
-function encodeRange(vals: DateRange): string {
-  // the list of val is encoded as YYYY-MM-DD and separated by a comma
-  // this happens after the "correction" of timestamps in the range selector
-  // that's why here is (yet again) some rounding to the start/end of the day.
-  const start =
-    vals[0] != null
-      ? moment(roundToMidnight(vals[0], "start"))?.format("YYYY-MM-DD")
-      : null;
-  const end =
-    vals[1] != null
-      ? moment(roundToMidnight(vals[1], "end"))?.format("YYYY-MM-DD")
-      : null;
-  if (start != null && end != null) {
-    return [start, end].join("_");
-  } else {
-    return "";
-  }
-}
-
-// the inverse of encodeRange
-function decodeRange(val: string): DateRange {
-  const vals = val.split("_");
-  const [start, end] = [startOfDay(vals[0]), endOfDay(vals[1])];
-  // if start and end are Date objects, return them
-  if (start instanceof Date && end instanceof Date) {
-    return [start, end];
-  } else {
-    return [undefined, undefined];
-  }
-}
-
-export function encodeFormValues(router: NextRouter, vals: any): void {
-  const { query } = router;
-  for (const key in vals) {
-    const val = vals[key];
-    if (key === "type" || key === "preset") continue; // we're already on the right page
-    if (val == null) {
-      delete query[key];
-    } else if (key === "range") {
-      query[key] = encodeRange(val);
-    } else {
-      query[key] = val;
-    }
-  }
-  router.replace({ query }, undefined, { shallow: true, scroll: false });
-}
-
-function decodeValue(val): boolean | number | string | DateRange {
-  if (val === "true") return true;
-  if (val === "false") return false;
-  const num = Number(val);
-  if (!isNaN(num)) return num;
-  return val;
-}
-
-/** a query looks like this:
- * user=academic&period=monthly&run_limit=1&member=true&uptime=short&cpu=1&ram=2&disk=3
- *
- * NOTE: the support for dedicated disk & vm does not work. the form is too complicated, not no need to support this yet.
- */
-export function decodeFormValues(
-  router: NextRouter,
-  type: "regular" | "boost" | "dedicated"
-): {
-  [key: string]: string | number | boolean;
-} {
-  const COMMON_FIELDS = ["user", "period", "range"] as const;
-  const FORM_FIELDS =
-    type === "dedicated"
-      ? [
-          ...COMMON_FIELDS,
-          "disk-size_gb",
-          "disk-name",
-          "disk-speed",
-          "vm-machine",
-        ]
-      : ([
-          ...COMMON_FIELDS,
-          "run_limit",
-          "member",
-          "uptime",
-          "cpu",
-          "ram",
-          "disk",
-        ] as const);
-
-  const minVal = type === "boost" ? 0 : 1;
-
-  const data = {};
-  for (const key in router.query) {
-    const val = router.query[key];
-    if (!(FORM_FIELDS as readonly string[]).includes(key)) continue;
-    if (typeof val !== "string") continue;
-    data[key] = key === "range" ? decodeRange(val) : decodeValue(val);
-  }
-
-  // we also have to sanitize the values
-  for (const key in data) {
-    const val = data[key];
-    switch (key) {
-      case "user":
-        if (!["academic", "commercial"].includes(val)) {
-          data[key] = "academic";
-        }
-        break;
-
-      case "period":
-        if (!["monthly", "yearly", "range"].includes(val)) {
-          data[key] = "monthly";
-        }
-        break;
-
-      case "range":
-        // check that val is an array of length 2 and both entries are Date objects
-        if (!Array.isArray(val) || val.length !== 2 || !val.every(isDate)) {
-          data[key] = [undefined, undefined];
-        }
-        break;
-
-      case "run_limit":
-        // check that val is a number and in the range of 1 to 1000
-        if (typeof val !== "number" || val < 1 || val > MAX_ALLOWED_RUN_LIMIT) {
-          data[key] = 1;
-        }
-        break;
-
-      case "member":
-        if (typeof val !== "boolean") {
-          data[key] = true;
-        }
-        break;
-
-      case "uptime":
-        if (!["short", "medium", "day", "always_running"].includes(val)) {
-          data[key] = "short";
-        }
-        break;
-
-      case "cpu":
-        if (typeof val !== "number" || val < minVal || val > MAX_DISK) {
-          data[key] = 1;
-        }
-        break;
-
-      case "ram":
-        if (typeof val !== "number" || val < minVal || val > MAX_GB_RAM) {
-          data[key] = 1;
-        }
-        break;
-
-      case "disk":
-        if (typeof val !== "number" || val < minVal || val > MAX_DISK) {
-          data[key] = 1;
-        }
-        break;
-
-      case "disk-size_gb":
-        if (typeof val !== "number" || !DEDICATED_DISK_SIZES.includes(val)) {
-          data[key] = DEFAULT_DEDICATED_DISK_SIZE;
-        }
-        break;
-
-      case "disk-name":
-        try {
-          testDedicatedDiskNameBasic(val);
-        } catch {
-          data[key] = "";
-        }
-        break;
-
-      case "disk-speed":
-        if (!DEDICATED_DISK_SPEEDS.includes(val)) {
-          data[key] = DEFAULT_DEDICATED_DISK_SPEED;
-        }
-        break;
-
-      case "vm-machine":
-        if (PRICES.vms[val] == null) {
-          data[key] = DEFAULT_DEDICATED_VM_MACHINE;
-        }
-        break;
-
-      default:
-        console.log(`decodingFormValues: unknown key '${key}'`);
-        delete data[key];
-    }
-  }
-
-  // hosting quality vs. uptime restriction:
-  if (["always_running", "day"].includes(data["uptime"])) {
-    data["member"] = true;
-  }
-
-  if (type === "dedicated") {
-    data["type"] = data["vm-machine"] != null ? "vm" : null;
-
-    // if any key in data starts with "disk-" then set data["type"] to "disk"
-    if (data["type"] == null) {
-      for (const key in data) {
-        if (key.startsWith("disk-")) {
-          data["type"] = "disk";
-          break;
-        }
-      }
-    }
-
-    if (data["type"] === "disk") {
-      data["period"] = "monthly";
-    }
-    if (data["type"] === "vm") {
-      data["period"] = "range";
-    }
-  }
-
-  return data;
 }
